requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
Django==3.0.6
django-crum==0.7.9
django-mssql-backend==2.8.1
django-widget-tweaks==1.4.12
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.5.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pyodbc==4.0.39
pypdf==3.9.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.2
webencodings==0.5.1
xhtml2pdf==0.2.11
